feat: refatorando activitys
