+++ app/src/main/java/br/luizfilipe/orgs/ui/activity/user/CadastroUserActivity.kt
@@ -1,13 +1,14 @@
 package br.luizfilipe.orgs.ui.activity.user
 
 import android.os.Bundle
+import android.util.Log
 import android.view.View
-import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import br.luizfilipe.orgs.database.AppDataBase
 import br.luizfilipe.orgs.databinding.ActivityCadastroUserBinding
 import br.luizfilipe.orgs.extensions.toHash
+import br.luizfilipe.orgs.extensions.toast
 import br.luizfilipe.orgs.model.User
 import kotlinx.coroutines.launch
 
@@ -19,6 +20,7 @@ class CadastroUserActivity : AppCompatActivity() {
         val db = AppDataBase.getInstance(this)
         db.userDaoRoom()
     }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
@@ -29,42 +31,25 @@ class CadastroUserActivity : AppCompatActivity() {
     }
 
     private fun configuraBotaoSalvar() {
-        val buttonSalvar = binding.cadastroUserSalvar
-        buttonSalvar.setOnClickListener(View.OnClickListener {
-            val nome = binding.cadastroUserCampoNome.text.toString()
-            val email = binding.cadastroUserCampoEmail.text.toString()
-            val senha = binding.cadastroUserCampoSenha.text.toString()
-            val telefone = binding.cadastroUserCampoTelefone.text.toString()
-
-            if (validaInformacoes(nome, email, senha, telefone)) {
-                val user = criaUser()
-                try {
-                    lifecycleScope.launch {
-                        userDao.salva(user)
-                        finish()
-                    }
-                } catch (e: Exception) {
-                    Toast.makeText(
-                        this, "Erro ao criar usuario",
-                        Toast.LENGTH_SHORT
-                    ).show()
-                }
-            }else{
-                Toast.makeText(
-                    this, "Todos os campos devem ser preenchidos",
-                    Toast.LENGTH_SHORT
-                ).show()
+        binding.cadastroUserSalvar.setOnClickListener(View.OnClickListener {
+            val novoUsuario = criaUser()
+            lifecycleScope.launch {
+                cadastra(novoUsuario)
             }
         })
-
     }
 
-    private fun validaInformacoes(
-        nome: String,
-        email: String,
-        senha: String,
-        telefone: String
-    ) = nome.isNotEmpty() && email.isNotEmpty() && senha.isNotEmpty() && telefone.isNotEmpty()
+    private fun cadastra(usuario: User) {
+        lifecycleScope.launch {
+            try {
+                userDao.salva(usuario)
+                finish()
+            } catch (e: Exception) {
+                Log.e("CadastroUsuario", "configuraBotaoSalvar: ", e)
+                toast("Falha ao cadastrar usuario.")
+            }
+        }
+    }
 
     private fun criaUser(): User {
         val nome = binding.cadastroUserCampoNome.text.toString()
